Backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=24.0",
    "drf-spectacular>=0.27.0",
    "python-decouple>=3.8",
    "qrcode[pil]>=7.4.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "dj-database-url>=3.0.1",
    "drf-nested-routers>=0.94.2",
]
